better url checking using apache commons-validator refactoring
